ci(changesets): version packages (#1473)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to dev, this PR will
be updated.


# Releases
## vue-docgen-cli@4.57.1

### Patch Changes

-
[`00f0a5b6`](https://github.com/vue-styleguidist/vue-styleguidist/commit/00f0a5b6b6f29ac32234f45e62554eb7a1abf824)
Thanks [@elevatebart](https://github.com/elevatebart)! - fix
templates.expose is not a function

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>