ci(changesets): version packages (#1489)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to dev, this PR will
be updated.


# Releases
## vue-docgen-cli@4.59.0

### Minor Changes

-
[`40cdd093`](https://github.com/vue-styleguidist/vue-styleguidist/commit/40cdd0931c580a8b3224504a582aab8a818d00c4)
Thanks [@elevatebart](https://github.com/elevatebart)! - protect md
input that is not string

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>