ci(changesets): version packages (#1497)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to dev, this PR will
be updated.


# Releases
## vue-docgen-cli@4.61.2

### Patch Changes

-
[`d1cbd701`](https://github.com/vue-styleguidist/vue-styleguidist/commit/d1cbd701b21794e416b7f099e95a5b322999c382)
Thanks [@elevatebart](https://github.com/elevatebart)! - make ignore
work with the watcher

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>