[vue-docgen-cli]: Make `extractConfig` async to allow importing module js config files
Using the dynamic `import()` operation allows to import commonJS and ESM files,
but requires the function to be async.
To allow `import()` within commonJS files Typescript requires the `moduleResolution` to
be set to `node16`, so we need to set our node dependency to node 16+.
This also allows us to set the code target to ES2021 supported by node 16, resulting in
smaller distribution size.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>