Add TypeScript generic type parameters to event payloads (#1538)
This PR adds generic type parameters to event payloads if specified in
TypeScript.

If this is a feature that was deliberatly not implemented so far, or is
not supposed to be added at all, just let me know and we will implement
it within our project (I suppose we can just add a parse handler). If
not this PR can maybe lay the foundation for supporting these types.

Our use case is mainly stemming from event payloads being arrays.
`vue-docgen-api` extracts these types currently just as `Array`, which
is not helpful enough. With this PR, the extraction would look like
this:

```json
{"name": "store-error", "type": {"names": ["Array"], "elements": [{"name": "Error"}]}},
```

instead of just

```json
{"name": "store-error", "type": {"names": ["Array"]}},
```

---------

Co-authored-by: Barthélémy Ledoux <ledouxb@gmail.com>