ci(changesets): version packages (#1574)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to dev, this PR will
be updated.


# Releases
## vue-docgen-api@4.72.5

### Patch Changes

-
[#1573](https://github.com/vue-styleguidist/vue-styleguidist/pull/1573)
[`d4c2deca`](https://github.com/vue-styleguidist/vue-styleguidist/commit/d4c2decae791b7311ea1c0988aa3935501ea4024)
Thanks [@thibaudszy](https://github.com/thibaudszy)! - improve
performance of the path resolution

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>