ci(changesets): version packages (#1577)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to dev, this PR will
be updated.


# Releases
## vue-docgen-api@4.73.0

### Minor Changes

-
[#1576](https://github.com/vue-styleguidist/vue-styleguidist/pull/1576)
[`3f1d9216`](https://github.com/vue-styleguidist/vue-styleguidist/commit/3f1d92161ae409cbcd9f3f7794110cfa3e8f115a)
Thanks [@elevatebart](https://github.com/elevatebart)! - parse the
defineOptions object for values

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>