fix: Include setupSlotHandler in the final build (#1624)
This pull request addresses the issue where the setupSlotHandler was not
included in the final build of the vue-styleguidist. The bug caused the
functionality of setupSlotHandler to be missing in the compiled output.

With this fix, the setupSlotHandler functionality is now correctly
included and available in the final build.

Please review this pull request and consider merging it into the main
branch to incorporate the fix.