ci(changesets): version packages (#1625)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to dev, this PR will
be updated.


# Releases
## vue-docgen-api@4.75.1

### Patch Changes

-
[#1624](https://github.com/vue-styleguidist/vue-styleguidist/pull/1624)
[`bc4ff434`](https://github.com/vue-styleguidist/vue-styleguidist/commit/bc4ff434c4fa15df4ca7bc3dc982b86738b23ff9)
Thanks [@azhirov](https://github.com/azhirov)! - Include
setupSlotHandler in the final build

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>