ci(changesets): version packages (#1648)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to dev, this PR will
be updated.


# Releases
## vue-docgen-cli@4.77.0

### Minor Changes

-
[`ab2b90a6`](https://github.com/vue-styleguidist/vue-styleguidist/commit/ab2b90a6b566964feae09b413a14e6008a1bb5a2)
Thanks [@elevatebart](https://github.com/elevatebart)! - add a param to
all templates to access the full doc

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>